Difficulty and Extra Lives Tweak
The "normal" difficulty for now is 10 waves instead of 15. The ptero spawn rate is a little more challenging. To speed up the pace and balance extra lives, the lives will automatically replenish health when the player's health is depleted. There will no longer be a continue prompt. FIXED stage complete audio not queuing properly.